refactor Ht* components into module
