+++ client/src/components/Ht/HtCard/index.ts
@@ -1,0 +1,3 @@
+import HtCard from 'components/Ht/HtCard/HtCard';
+
+export default HtCard;
