+++ client/src/components/Ht/HtDropdown/index.ts
@@ -1,0 +1,3 @@
+import HtDropdown from 'components/Ht/HtDropdown/HtDropdown';
+
+export default HtDropdown;
