+++ client/src/components/Ht/HtModal/index.ts
@@ -1,0 +1,3 @@
+import HtModal from 'components/Ht/HtModal/HtModal';
+
+export default HtModal;
