+++ client/src/components/Ht/HtSpinner/index.ts
@@ -1,0 +1,3 @@
+import HtSpinner from 'components/Ht/HtSpinner/HtSpinner';
+
+export default HtSpinner;
