+++ client/src/components/Ht/HtTooltip/index.ts
@@ -1,0 +1,3 @@
+import HtTooltip from 'components/Ht/HtTooltip/HtTooltip';
+
+export default HtTooltip;
