--- client/src/components/HtCard/index.ts
@@ -1,3 +1,0 @@
-import HtCard from './HtCard';
-
-export default HtCard;
