--- client/src/components/HtDropdown/index.ts
@@ -1,3 +1,0 @@
-import HtDropdown from './HtDropdown';
-
-export default HtDropdown;
