--- client/src/components/HtModal/index.ts
@@ -1,3 +1,0 @@
-import HtModal from './HtModal';
-
-export default HtModal;
