--- client/src/components/HtSpinner/index.ts
@@ -1,3 +1,0 @@
-import HtSpinner from './HtSpinner';
-
-export default HtSpinner;
