--- client/src/components/HtTooltip/index.ts
@@ -1,3 +1,0 @@
-import HtTooltip from './HtTooltip';
-
-export default HtTooltip;
